+++ src/ambition/version_details.cc
@@ -21,6 +21,7 @@ std::vector<VersionUpdates> versionDetails = {
         .items = {
           _("Added a clone button to Caravans.  If there is an idle Caravan, it is used as the clone.  Otherwise, a new Caravan spawns at the Market closest to the Caravan's first stop."),
           _("Added a caravan clone button to the trade report, next to the paste route button."),
+          _("Automatically \"heal\" caravans by replacing damaged ones using the above caravan cloning feature.  Caravans heading to a Market will be replaced if they have any damage, and caravans which do not have any Market stops will be replaced if they fall to 80% or below."),
           _("Show the skill icon of the unit being trained in Towns under the progress bar on the world map and in the info pane."),
         },
       },